Preserve selection on paste
We do this by parsing pasted content that is not native CQL HTML generated by ProseMirror (e.g. <query-str> etc.) on a separate path. Plain HTML is parsed into a plain string, which if left to the usual parse behaviour would enter the document as a single queryStr node, and then be transformed into a richer representation in `applyQueryToTr`.

This transformation has the side-effect of occasionally producing an incorrect selection state, because the
caret position in a query-str node, denoted by $ (e.g. <query-str>tag:$</query-str>, pos 5) will not be correct when that string is parsed into the richer document (e.g. <query-str></query-str><chip><chip-key>t$ag</chip-key><chip-value></chip-value></chip>.)

So when we receive text, or HTML that we think is text, we prevent the default paste behaviour, and instead parse the input into a CQL document and run a diff update, as we would if the `value` attribute was updated on the component.

Another way to achieve this without custom paste behaviour might be to preserve selection across the transformation from plain text (<query-str>tag:</query-str>) to rich query. This is a little fiddlier, because we must account for the fact that prefixes are optional, so a mapping step to normalise the selection to account for prefixes would be necessary before we mapped the selection across the PM document.