Refactor: extract debug rendering from CqlInput
Refactor the rendering of debug information so that it is all handled
based on events emitted from the CqlInput component, rather than being
part of the component code itself.

Cleaner separation of concerns and a smaller public API for the
component.

One trade-off is that we had to remove the suggestions debug output for
the time being, because suggestion info is not emitted, and would have
to be fetched separately. Something to think about for future work
though: is there a good public API that we could provide which would
facilitate this?

Co-authored-by: Jonathon Herbert <jonathon.herbert@guardian.co.uk>